[#656] turn #create_or_update_external_link into a concern
- include module with business person
- abstract test into rspec support module
