+++ spec/models/business_person_spec.rb
@@ -1,5 +1,6 @@
 require 'rails_helper'
 
-describe BusinessPerson do
+describe BusinessPerson, :external_link, :type => :model do
   it { is_expected.to have_db_column(:sec_cik) }
+  create_or_update_external_link_test('business_person', 'person')
 end
